Add basic routing and view logic to home app
The main urls.py of the project (shlifim/urls.py) is redirecting
all the routing to home/urls.py.
And the home/urls.py currently holds routes only for the about
page and landingpage.
Correspondingly, there are 2 view functions: about and landingpage
which just renders and returns the HTML file.