Added link from tags to explore page
In every place we display tags, (explore page and display question
page) I added an option to click on the tag's badge and it will redirect
to the explore page filtered to show all question with this tag.

No tests were added because this link's url has allready been
tested during the tags page feature.